pyproject.toml
[tool.poetry]
name = "p-v-monster-translate"
version = "0.1.0"
description = "My translate project."
authors = ["Wei Hung <wayne930242@gmail.com>"]
license = "All rights reserved."
readme = "README.md"
packages = [{ include = "my_tm", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
markdown = "^3.6"
colorlog = "^6.8.2"
langchain = "^0.2.1"
langchain-text-splitters = "^0.2.0"
langchain-openai = "^0.1.20"
inflect = "^7.3.1"
langgraph = "^0.2.28"
markdownify = "^0.13.1"
ebooklib = "^0.18"
python-docx = "^1.1.2"
fpdf = "^1.7.2"
weasyprint = "^62.3"
pypdf2 = "^3.0.1"

[tool.poetry.dev-dependencies]
mypy = "^1.10.0"

[tool.poetry.scripts]
my-tm = "my_tm.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
